// src/app/interfaces/ocr-result.ts
export interface OcrResult {
    clientCompany: string;
    userName: string;
    idCompanyNo: string;
    unitType: string;
    unitSerialNo: string;
    vehicleType: string;
    regNo: string;
    vinNo: string;
    vehicleDescription: string;
    serviceCentre: string;
    serviceCentreTel: string;
    authorisedTechnicianName: string;
    date: string;
  }
  